move Stake and StakeDistribution to Primitive.Types
nix patches
