+++ lib/core/cardano-wallet-core.cabal
@@ -63,6 +63,7 @@ library
     , persistent-template
     , random
     , retry
+    , scientific
     , servant
     , servant-server
     , split
