+++ lib/core/src/Cardano/Wallet/Primitive/Types.hs
@@ -105,6 +105,8 @@ module Cardano.Wallet.Primitive.Types
 
     -- * Stake Pools
     , PoolId(..)
+    , Stake (..)
+    , StakeDistribution (..)
     , poolIdBytesLength
 
     -- * Querying
@@ -148,6 +150,8 @@ import Crypto.Number.Generate
     ( generateBetween )
 import Crypto.Random.Types
     ( MonadRandom )
+import Data.Aeson
+    ( FromJSON (..), Value (..), defaultOptions, genericParseJSON )
 import Data.Bifunctor
     ( bimap )
 import Data.ByteArray.Encoding
@@ -166,6 +170,8 @@ import Data.Proxy
     ( Proxy (..) )
 import Data.Quantity
     ( Percentage, Quantity (..) )
+import Data.Scientific
+    ( toBoundedInteger )
 import Data.Set
     ( Set )
 import Data.String
@@ -510,6 +516,32 @@ instance FromText PoolId where
                  <> show (BS.length bytes)
         where
             textDecodingError = Left . TextDecodingError . show
+
+-- | Coins are stored as Lovelace (reminder: 1 Lovelace = 1e-6 ADA)
+newtype Stake = Stake { getStake :: Word64 } deriving (Show, Eq, Generic)
+
+data StakeDistribution = StakeDistribution
+    { dangling :: Stake
+    , pools :: [(PoolId,Stake)]
+    , unassigned :: Stake
+    } deriving (Eq, Show, Generic)
+
+instance FromJSON StakeDistribution where
+    parseJSON = genericParseJSON defaultOptions
+
+instance FromJSON Stake where
+    parseJSON (Number n) = case toBoundedInteger n of
+        Nothing ->
+            fail "stake should be non-negative integer"
+        Just stake ->
+            return (Stake stake)
+    parseJSON _ = fail "stake should be numeric"
+
+instance FromJSON PoolId where
+    parseJSON (String txt) = case fromText txt of
+        Left (TextDecodingError err) -> fail err
+        Right bs -> return bs
+    parseJSON _ = fail "stake pool id should be text"
 
 {-------------------------------------------------------------------------------
                                     Block
